requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==44.0.2
ifaddr==0.2.0
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
pyserial==3.5
setuptools==78.1.0
